lstm-trader/requirements.txt
# 🧠 LSTM Forex Predictor - Python Dependencies
# Real neural network training and prediction

# Core ML libraries
tensorflow>=2.13.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# API
Flask>=2.3.0
Flask-Cors>=4.0.0
gunicorn>=21.2.0

# Database
psycopg2-binary>=2.9.0

# Data fetching
requests>=2.31.0
yfinance>=0.2.0

# Visualization (optional)
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0

# Development
jupyter>=1.0.0
ipython>=8.0.0
